Cambios modelos

+++ laravel/app/Http/Controllers/FileController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\File;
 use App\Models\Place;
+use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\DB;
@@ -171,8 +172,9 @@ class FileController extends Controller
     public function destroy(File $file)
     {
         //
+        $post = Post::where('file_id', $file->id)->first();
         $place = Place::where('file_id', $file->id)->first();
-        if (is_null($place)){
+        if (is_null($place) && is_null($post)){
             if (\Storage::disk('public')->exists($file->filepath)) {
                 File::destroy($file->id);
                 \Storage::disk('public')->delete($file->filepath);
